miApiJWT/requirements.txt
uvicorn[standard]
fastapi 
python-jose[cryptography]
python-multipart